requirements.txt
Django==4.2.23
djangorestframework==3.15.2
django-cors-headers==4.3.1
qrcode==8.2
Pillow==10.1.0
python-docx==0.8.11
reportlab==4.0.7
